rename files, regen abi, update docs
