+++ contracts/README.md
@@ -3,8 +3,6 @@ This directory contains Solidity contracts implementing the Teleporter messaging
 
 This directory is set up as a [Foundry](https://github.com/foundry-rs/foundry) project. Use the `scripts/install_foundry.sh` to install the correct version of the ava-labs fork of foundry. Further documentation about given contracts can be found in`src/Teleporter/`.
 
-Other folders include other cross-chain applications that don't necessarily directly interact with the Teleporter contract. E.g. `src/OffChainMessageContracts`. Additional documentation for those can be found in their respective folders.
-
 ## Building and Running
 - To compile the contracts run `forge build` from this directory.
 - Similarly, to run unit tests, run `forge test`.
